json handling

+++ CycleSafeJava/app/src/main/java/com/example/cyclesafejava/data/Statistics.java
@@ -1,7 +1,7 @@
 package com.example.cyclesafejava.data;
 
 public class Statistics {
-    public float LongestRide;
-    public float FastestSpeed;
-    public float TotalDistance;
+    public Double LongestRide;
+    public Double FastestSpeed;
+    public Double TotalDistance;
 }
